pyproject.toml
[tool.poetry]
name = "discord-phone-proxy"
version = "0.1.0"
description = ""
authors = ["Ben McKibben <dev@trash.house>"]

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.19.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
aiohttp = "^3.6.0"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
websockets = "^10.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
